Extract tab screens to its own components
